+++ src/main/java/gtclassic/api/gui/GTGuiCompSort.java
@@ -6,19 +6,21 @@ import java.util.List;
 import ic2.core.block.base.tile.TileEntityMachine;
 import ic2.core.inventory.gui.GuiIC2;
 import ic2.core.inventory.gui.components.GuiComponent;
-import ic2.core.platform.registry.Ic2GuiComp;
+import ic2.core.util.math.Box2D;
 import net.minecraft.client.gui.GuiButton;
+import net.minecraft.client.resources.I18n;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class GTGuiCompSort extends GuiComponent {
 
 	TileEntityMachine block;
+	// private static final Box2D BOX = new Box2D(7, 62, 18, 18);
 	int x;
 	int y;
 
 	public GTGuiCompSort(TileEntityMachine tile, int x, int y) {
-		super(Ic2GuiComp.nullBox);
+		super(new Box2D(x, y, 18, 18));
 		this.block = tile;
 		this.x = x;
 		this.y = y;
@@ -26,7 +28,7 @@ public class GTGuiCompSort extends GuiComponent {
 
 	@Override
 	public List<ActionRequest> getNeededRequests() {
-		return Arrays.asList(ActionRequest.GuiInit, ActionRequest.ButtonNotify);
+		return Arrays.asList(ActionRequest.GuiInit, ActionRequest.ButtonNotify, ActionRequest.ToolTip);
 	}
 
 	@Override
@@ -40,6 +42,14 @@ public class GTGuiCompSort extends GuiComponent {
 	public void onButtonClick(GuiIC2 gui, GuiButton button) {
 		if (button.id == 1) {
 			this.block.getNetwork().initiateClientTileEntityEvent(this.block, 1);
+		}
+	}
+
+	@Override
+	@SideOnly(Side.CLIENT)
+	public void onToolTipCollecting(GuiIC2 gui, int mouseX, int mouseY, List<String> tooltips) {
+		if (this.isMouseOver(mouseX, mouseY)) {
+			tooltips.add(I18n.format("button.sort"));
 		}
 	}