+++ src/main/java/gtclassic/common/gui/GTGuiCompUUMAssembler.java
@@ -4,25 +4,28 @@ import java.awt.Color;
 import java.util.Arrays;
 import java.util.List;
 
+import gtclassic.api.helpers.GTHelperMath;
 import gtclassic.common.tile.GTTileUUMAssembler;
 import ic2.core.inventory.gui.GuiIC2;
 import ic2.core.inventory.gui.components.GuiComponent;
-import ic2.core.platform.registry.Ic2GuiComp;
+import ic2.core.util.math.Box2D;
+import net.minecraft.client.resources.I18n;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class GTGuiCompUUMAssembler extends GuiComponent {
 
 	GTTileUUMAssembler block;
+	private static final Box2D BOX = new Box2D(97, 57, 57, 18);
 
 	public GTGuiCompUUMAssembler(GTTileUUMAssembler tile) {
-		super(Ic2GuiComp.nullBox);
+		super(BOX);
 		this.block = tile;
 	}
 
 	@Override
 	public List<ActionRequest> getNeededRequests() {
-		return Arrays.asList(ActionRequest.FrontgroundDraw);
+		return Arrays.asList(ActionRequest.FrontgroundDraw, ActionRequest.ToolTip);
 	}
 
 	@Override
@@ -35,5 +38,21 @@ public class GTGuiCompUUMAssembler extends GuiComponent {
 		gui.drawString("Cost: " + cost, 12, 38, Color.cyan.hashCode());
 		int per = this.block.getAmountPer();
 		gui.drawString("Produced: " + per, 12, 48, Color.cyan.hashCode());
+	}
+
+	@Override
+	@SideOnly(Side.CLIENT)
+	public void onToolTipCollecting(GuiIC2 gui, int mouseX, int mouseY, List<String> tooltips) {
+		if (this.isMouseOver(mouseX, mouseY)) {
+			if (mouseX < 114) {
+				tooltips.add(I18n.format("button.uum0"));
+			}
+			if (GTHelperMath.within(mouseX, 115, 130)) {
+				tooltips.add(I18n.format("button.uum1"));
+			}
+			if (mouseX > 130 && this.block.inventory.get(11).isEmpty()) {
+				tooltips.add(I18n.format("button.uum2"));
+			}
+		}
 	}
 }