+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added UU Matter Plasma as a dynamite magic fuel!
 - Fixed some recipes not using ender chest ore dict
 - Fixed buffers not checking the correct stack size to see if an inventory was full, buffers now output redstone if there no empty slots
+- Added mouseover tooltips for every single GTC Gui that has a button.
 
 # 1.1.1
 - We wrapping this project up on 1.12.2, no data net, no massive open ended time sink projects. Couple more feature additions and then GTC 1.12.2 is done except bug fixes and QOL.